switching to unsigned integers in c solutions to allow for some optimizations with regards to division by constants

+++ c/src/p002.c
@@ -2,24 +2,24 @@
 #include <limits.h>
 
 #if (UINT_MAX >= 0xffffffff)
-typedef int  int32_t;
+typedef unsigned int  uint32_t;
 #else
-typedef long int32_t;
+typedef unsigned long uint32_t;
 #endif
 
 #define true (1)
 
 int main(void) {
-    int32_t limit = 4000000;
-    int32_t a1 = 0;
-    int32_t a2 = 2;
-    int32_t sum = 2;
+    uint32_t limit = 4000000;
+    uint32_t a1 = 0;
+    uint32_t a2 = 2;
+    uint32_t sum = 2;
 
     while (true) {
         /* Every third term is even
         ** f(n) = 4*f(n-3) + f(n-6)
         */
-        int32_t a3 = 4*a2 + a1;
+        uint32_t a3 = 4*a2 + a1;
         a1 = a2;
         a2 = a3;
         if (a3 > limit) {
@@ -27,5 +27,5 @@ int main(void) {
         }
         sum += a3;
     }
-    printf("%d\n", sum);
+    printf("%lu\n", (unsigned long)sum);
 }
